Storybook: Fix missing background path in NTP
