--- components/brave_new_tab_ui/components/default/exchange/assets/exchange-logo.tsx
@@ -1,16 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
-* License, v. 2.0. If a copy of the MPL was not distributed with this file,
-* You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-import * as React from 'react'
-
-export default class ExchangeLogo extends React.PureComponent {
-
-  render () {
-    return (
-      <svg width={'32'} height={'32'} xmlns={'http://www.w3.org/2000/svg'}>
-        <path d={'M20.64 17.934l2.706 2.706L16 27.985 8.653 20.64l2.707-2.706L16 22.572l4.64-4.637zM6.72 13.293L9.431 16l-2.71 2.708L4.014 16l2.707-2.707zm9.278 0L18.707 16 16 18.707 13.293 16 16 13.293zm9.281 0L27.986 16l-2.704 2.708L22.573 16l2.706-2.707zM16 4.016l7.346 7.346-2.706 2.707L16 9.428l-4.639 4.64-2.707-2.707L16 4.013z'} fill={'#F0B90B'} fillRule={'evenodd'} />
-      </svg>
-    )
-  }
-}
